django-chatbot/django_chatbot/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
Django==5.0.4
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.3
kombu==5.3.7
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numba==0.59.1
numpy==1.26.4
openai==1.23.6
openai-whisper==20231117
pillow==10.3.0
prompt-toolkit==3.0.43
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
regex==2024.4.16
requests==2.31.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
sympy==1.12
tiktoken==0.6.0
torch==2.3.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
